backend/requirements.txt
python-dotenv>=1.0,<2.0
django==5.2.5
djangorestframework>=3.15,<3.16
django-filter>=24.3,<25.0
django-cors-headers>=4.5,<5.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg[binary]>=3.2,<3.3
